Feat: using Spring data JPA #14
